Schematics As-Built Drafts
Disclaimer: Electrical Diagram is not reflective of the actual circuit but is functionally identical

Originally drafts of the project before assembly, then revised to as-built to be updated with changes made to the physical apparatus. The P&ID has an excess of information displayed for quick reference because I prefer it over searching through the Rock Candy Maker P&ID Spec Sheet. The Electrical Diagram and P&ID were made using AutoCAD Electrical and AutoCAD Plant 3D respectively